Move cleaners tp a separate directory
