+++ src/utilities/quiz_seasons.py
@@ -7,7 +7,7 @@ from typing import Dict, List
 
 import pandas as pd
 
-from .season_cleaner import SeasonCleaner
+from ..cleaners.season_cleaner import SeasonCleaner
 
 
 @dataclass
